requirements.txt
# setup packages
serial
time

# sonar packages
RPi.GPIO

#optical flow sensor packages
smbus2